backend/requirements.txt
# ============================================
# SAVIA Backend — Python Dependencies (Production)
# Only packages used by main.py, db_engine.py, auth.py,
# ai_engine.py, log_preprocessor.py, s3_storage.py
# ============================================

# FastAPI stack
fastapi==0.135.1
uvicorn==0.42.0
starlette==0.52.1
python-multipart==0.0.26
pydantic==2.12.5
pydantic_core==2.41.5

# Authentication
PyJWT==2.11.0
bcrypt==5.0.0
cryptography==46.0.4

# Database
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48

# Data
pandas==2.3.3
numpy==2.4.2
openpyxl==3.1.5
pyarrow==23.0.0

# AI / Google Gemini
google-genai==1.62.0
google-generativeai==0.8.6
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-auth==2.49.0.dev0
googleapis-common-protos==1.72.0
proto-plus==1.27.1
protobuf==5.29.6

# Storage (MinIO/S3)
boto3==1.38.0

# HTTP / Networking
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
urllib3==2.6.3

# PDF
fpdf2==2.8.5
PyMuPDF==1.26.7

# Scheduling
APScheduler==3.11.2

# Utils
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
typing_extensions==4.15.0
anyio==4.12.1
sniffio==1.3.1
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
click==8.3.1
packaging==26.0
six==1.17.0
xlsxwriter==3.2.9
RapidFuzz==3.14.3
tenacity==9.1.2
